requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.2
sqlparse==0.4.2
tzdata==2021.5
django-filter
django-bootstrap-form
django-fas
